use atomic operate the size

+++ libs/iavl/config/dynamic_config_okchain.go
@@ -2,7 +2,7 @@ package config
 
 type IDynamicConfig interface {
 	GetIavlCacheSize() int
-	GetIavlFSCacheSize() int
+	GetIavlFSCacheSize() int64
 }
 
 var DynamicConfig IDynamicConfig = MockDynamicConfig{}
@@ -18,6 +18,6 @@ func (d MockDynamicConfig) GetIavlCacheSize() int {
 	return 10000
 }
 
-func (d MockDynamicConfig) GetIavlFSCacheSize() int {
+func (d MockDynamicConfig) GetIavlFSCacheSize() int64 {
 	return 10000
 }
